+++ coders/pcx.c
@@ -520,8 +520,8 @@ static Image *ReadPCXImage(const ImageInfo *image_info,ExceptionInfo *exception)
                                 MagickArraySize(pcx_info.bytes_per_line,
                                                 pcx_info.planes));
     if ((0 == pcx_packets) ||
-        ((size_t) (pcx_info.bits_per_pixel*pcx_info.planes*image->columns) >
-         (pcx_packets*8U)))
+        (((size_t) pcx_info.bits_per_pixel*pcx_info.planes*image->columns) >
+         ((size_t) pcx_packets*8U)))
       ThrowPCXReaderException(CorruptImageError,ImproperImageHeader,image);
     pcx_pixels=MagickAllocateMemory(unsigned char *,pcx_packets);
     if (pcx_pixels == (unsigned char *) NULL)
@@ -604,7 +604,7 @@ static Image *ReadPCXImage(const ImageInfo *image_info,ExceptionInfo *exception)
                   256 color images have their color map at the end of the file.
                 */
                 pcx_info.colormap_signature=ReadBlobByte(image);
-                (void) ReadBlob(image,3*image->colors,(char *) pcx_colormap);
+                (void) ReadBlob(image, (size_t) 3*image->colors,(char *) pcx_colormap);
                 p=pcx_colormap;
                 for (i=0; i < image->colors; i++)
                 {
@@ -620,7 +620,7 @@ static Image *ReadPCXImage(const ImageInfo *image_info,ExceptionInfo *exception)
     */
     for (y=0; y < (long) image->rows; y++)
     {
-      p=pcx_pixels+((unsigned long) y*pcx_info.bytes_per_line*pcx_info.planes);
+      p=pcx_pixels+((size_t) y*pcx_info.bytes_per_line*pcx_info.planes);
       q=SetImagePixels(image,0,y,image->columns,1);
       if (q == (PixelPacket *) NULL)
         break;
